+++ agents/group16_agent/bidding_strategies/bidding_strategy.py
@@ -1,0 +1,57 @@
+from geniusweb.issuevalue.Bid import Bid
+from geniusweb.issuevalue.Domain import Domain
+from geniusweb.bidspace.AllBidsList import AllBidsList
+from time import time
+from random import randint
+
+
+class BiddingStrategy:
+
+    def __init__(self, profile, progress, opponent_model):
+        self._profile = profile
+        self._progress = progress
+        self._opponent_model = opponent_model
+
+    def find_bid(self) -> Bid:
+        # compose a list of all possible bids
+        domain = self._profile.getDomain()
+        all_bids = AllBidsList(domain)
+
+        best_bid_score = 0.0
+        best_bid = None
+
+        # take 500 attempts to find a bid according to a heuristic score
+        for _ in range(500):
+            bid = all_bids.get(randint(0, all_bids.size() - 1))
+            bid_score = self._score_bid(bid)
+            if bid_score > best_bid_score:
+                best_bid_score, best_bid = bid_score, bid
+
+        return best_bid
+
+    def _score_bid(self, bid: Bid, alpha: float = 0.95, eps: float = 0.1) -> float:
+        """Calculate heuristic score for a bid
+
+        Args:
+            bid (Bid): Bid to score
+            alpha (float, optional): Trade-off factor between self interested and
+                altruistic behaviour. Defaults to 0.95.
+            eps (float, optional): Time pressure factor, balances between conceding
+                and Boulware behaviour over time. Defaults to 0.1.
+
+        Returns:
+            float: score
+        """
+        progress = self._progress.get(time() * 1000)
+
+        our_utility = float(self._profile.getUtility(bid))
+
+        time_pressure = 1.0 - progress ** (1 / eps)
+        score = alpha * time_pressure * our_utility
+
+        if self._opponent_model is not None:
+            opponent_utility = self._opponent_model.get_predicted_utility(bid)
+            opponent_score = (1.0 - alpha * time_pressure) * opponent_utility
+            score += opponent_score
+
+        return score
